Fix CA1028
See https://docs.microsoft.com/en-us/visualstudio/code-quality/ca1028-enum-storage-should-be-int32?view=vs-2017

+++ de4dot.blocks/cflow/Value.cs
@@ -18,7 +18,7 @@
 */
 
 namespace de4dot.blocks.cflow {
-	public enum ValueType : byte {
+	public enum ValueType : int {
 		Unknown,
 		Null,
 		Object,
